move cli functions to its own class
